Дополнение запроса при регистрации

+++ NAuthAPI/Account.cs
@@ -8,9 +8,9 @@ namespace NAuthAPI
         public string Hash { get; set; }
         public string Salt { get; set; }
         public bool IsBlocked { get; set; }
-        public ushort Attempts { get; set; }
+        public byte Attempts { get; set; }
 
-        public Account(ClaimsIdentity identity, string hash, string salt, bool isBlocked, ushort attempts)
+        public Account(ClaimsIdentity identity, string hash, string salt, bool isBlocked, byte attempts)
         {
             Identity = identity;
             Hash = hash;
